Organize 1D-specific reservoir computing code (#2161)
