+++ projects/reservoir/kuramoto-sivashinsky/configs/hybrid_train_config.yaml
