+++ projects/reservoir/kuramoto-sivashinsky/configs/ks_config.yaml
