+++ projects/reservoir/kuramoto-sivashinsky/configs/train_config.yaml
